separate incentive migration to new file
